+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## Unreleased
+### Changed
+* Use `gunzip-maybe` for s3 read streams
+
 ## [3.0.0-alpha.1] - 2016-01-19
 ### Fixed
 * Add files missing from build